get_storage_html: Introduced show_browse parameter
Parameter show_browse:
Displays an eye symbol right to the select box to open the current path in LoxBerry's File Manager. The eye symbol is disabled if the base path is outside of /opt/loxberry, as this is not allowed in LoxBerry's File Manager. (available from LoxBerry 2.0.0.4)

Feature for https://github.com/christianTF/LoxBerry-Plugin-LoxberryBackup/issues/22