logmanager.cgi: Fixes performance because of attention messages
Log database saves all message with severity warning and above as column in the database. logmanager.cgi shows an info symbol in the overview, with a popup of the collected messages.
With an huge amount of lines, logmanager.cgi is still quite fast (<400ms), but the html and jQuery needs several seconds to parse the content.

- Introduced a limit of 40 messages (this is around one screen height) to show only the last 40 messages.

Fixes #1107